requirements.txt
crewai
qdrant-client
supabase
cipher
n8n-python-sdk
python-dotenv
fastapi
uvicorn